Tie in world id logic to LocationUseCase (#96)

+++ Source/Python/application/use_case/location_use_cases.py
@@ -12,36 +12,44 @@ class LocationUseCase:
     _location_repository: LocationRepository
     _event_repository: EventRepository
 
-    def __init__(self, location_repository: LocationRepository, event_repository: EventRepository) -> None:
+    def __init__(
+            self, world_repository: WorldRepository, location_repository: LocationRepository, event_repository: EventRepository
+    ) -> None:
         if not isinstance(location_repository, LocationRepository):
             raise TypeError(f"Argument 'location_repository' must be of type {LocationRepository}")
         if not isinstance(event_repository, EventRepository):
             raise TypeError(f"Argument 'event_repository' must be of type {EventRepository}")
-
+        self._world_repository = world_repository
         self._location_repository = location_repository
         self._event_repository = event_repository
 
     @requires_authentication()
     def create(self, world_id: PrefixedUUID, **kwargs) -> Location:
+        self._validate_world_exists(world_id)
         kwargs["id"] = generate_prefixed_id("location")
         location = Location(**kwargs)
 
         self._location_repository.save(location)
+        self._world_repository.associate(world_id, location_id=location.id)
 
         return location
 
     @requires_authentication()
     def retrieve(self, world_id: PrefixedUUID, location_id: PrefixedUUID) -> Location:
+        self._validate_world_exists(world_id)
         if not location_id.prefix == "location":
             raise ValueError("Argument 'location_id' must be prefixed with 'location'")
+        associated_locations = self._world_repository.get_all_associated(world_id, locations=True)
+        if location_id not in associated_locations:
+            raise NameError(f"No location '{location_id}' is exists for world '{world_id}'")
 
         return self._location_repository.retrieve(location_id)
 
     @requires_authentication()
     def retrieve_all(self, world_id: PrefixedUUID, **kwargs) -> Set[Location]:
-
-
-        all_locations = self._location_repository.retrieve_all()
+        self._validate_world_exists(world_id)
+        associated_locations = self._world_repository.get_all_associated(world_id, locations=True)
+        all_locations = {location for location in self._location_repository.retrieve_all() if location.id in associated_locations}
         name_filtered_locations, kwargs = FilteringUseCase.filter_named_entities(all_locations, **kwargs)
         tag_filtered_locations, kwargs = FilteringUseCase.filter_tagged_entities(name_filtered_locations, **kwargs)
         span_filtered_locations, kwargs = FilteringUseCase.filter_spanning_entities(tag_filtered_locations, **kwargs)
@@ -52,6 +60,11 @@ class LocationUseCase:
 
     @requires_authentication()
     def update(self, world_id: PrefixedUUID, location: Location) -> None:
+        self._validate_world_exists(world_id)
+        associated_locations = self._world_repository.get_all_associated(world_id, locations=True)
+        if location.id not in associated_locations:
+            raise NameError(f"No location '{location.id}' is exists for world '{world_id}'")
+
         self._location_repository.retrieve(location.id)
         self._validate_linked_events_still_intersect_for_update(location)
 
@@ -59,12 +72,21 @@ class LocationUseCase:
 
     @requires_authentication()
     def delete(self, world_id: PrefixedUUID, location_id: PrefixedUUID) -> None:
+        self._validate_world_exists(world_id)
         if not location_id.prefix == "location":
             raise ValueError("Argument 'location_id' must be prefixed with 'location'")
+        associated_locations = self._world_repository.get_all_associated(world_id, locations=True)
+        if location_id not in associated_locations:
+            raise NameError(f"No location '{location_id}' is exists for world '{world_id}'")
 
         self._validate_no_linked_events_for_delete(location_id)
 
         return self._location_repository.delete(location_id)
+
+    def _validate_world_exists(self, world_id: PrefixedUUID) -> None:
+        if not world_id.prefix == "world":
+            raise ValueError("Argument 'world_id' must be prefixed with 'world'")
+        self._world_repository.retrieve(world_id)
 
     def _validate_no_linked_events_for_delete(self, location_id: PrefixedUUID) -> None:
         linked_event_ids = [str(event.id) for event in self._event_repository.retrieve_all(location_id=location_id)]
